v1.3:
- Drop support for old versions of Python
- Returning variable is now Schema object
