+++ .travis.yml
@@ -1,11 +1,9 @@
 language: python
 python:
-  - "2.7"
-  - "3.3"
-  - "3.4"
   - "3.5"
-  - "3.5-dev" # 3.5 development branch
-  - "nightly" # currently points to 3.6-dev
+  - "3.6"
+  - "3.7"
+  - "3.8"
 # command to install dependencies
 install: "pip install -r test_requirements.txt"
 # command to run tests
