+++ LICENSE
@@ -1,4 +1,4 @@
-Copyright (c) 2015, Vladimir Chub
+Copyright (c) 2020, Vladimir Chub
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
